Fix numerous windows warnings through judicious use of size_t.
git-svn-id: svn+ssh://svn.code.sf.net/p/lemur/code/indri/trunk@1255 14020d50-3050-45fe-8e0c-5ed7005ca764

+++ include/indri/TagList.hpp
@@ -111,7 +111,7 @@ namespace indri
         // look through the tags vector; they're already in sorted order by open
         // position.  Only add closed tags.
 
-        for( unsigned int i=0; i<_tags.size(); i++ ) {
+        for( size_t i=0; i<_tags.size(); i++ ) {
           tag_entry& entry = _tags[i];
 
           if( entry.end >= 0 ) {// data field might be empty at head of doc
@@ -143,7 +143,7 @@ namespace indri
       // in this case, we'll treat the list of tags in this list
       // as if they were offsets into a metadata list
       void writeMetadataList( indri::utility::greedy_vector<MetadataPair>& pairs, indri::utility::Buffer& buffer, const char* docText ) {
-        for( unsigned int i=0; i<_tags.size(); i++ ) {
+        for( size_t i=0; i<_tags.size(); i++ ) {
           tag_entry& entry = _tags[i];
 
           if( entry.end > 0 ) {
